requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
langchain==0.0.335
langchain-ollama==0.1.3  
openai==1.3.5
PyPDF2==3.0.1
tiktoken==0.5.1
pymupdf==1.23.8 
faiss-cpu==1.7.4
PyPDF2
rank_bm25